refactor(frontend): extract right-panel basic/columns/access panes (P2-2 F17d)
The three largest tab panes move out of DataStudioRightPanel.vue into
self-sufficient children that inject dataStudioCtx directly (provided by
DataStudioNew, so no new wiring): DataStudioRightPanelBasic (meta
editing, Doris config, trend triggers — TableTrendDialog moves in with
its sole callers), DataStudioRightPanelColumns (field definitions
editing), DataStudioRightPanelAccess (metrics grid, trend/top-user
tables). DDL and versions panes are small and stay in the parent.

CSS ownership follows the F16c rule: pane-exclusive rules move with
their pane; scaffolding shared with the remaining parent panes
(.meta-section, .section-block/-fill/-header/-title/-actions,
.meta-scroll) converts to .meta-tabs :deep(...) in the parent so one
copy styles both sides. Verified no conflicting property overlap on the
only specificity-tied selector (.section-header.small).

DataStudioRightPanel.vue: 1299 -> 625 lines — under the P2-2 <800
target. New children: Basic 360 / Columns 243 / Access 163 lines.
The mount smoke gains a per-pane case (each child mounts with a fake
ctx and renders its marker node).

Verified: npm run lint (0 errors), npm run test (31 files / 160 passed),
npm run build; demo-mode Playwright smoke (11 checks): all five panes
render with data (5 field rows, 6 metric cards, real demo DDL), trend
dialog opens from the Basic child with an ECharts canvas, vertical
resize intact, and a computed-style probe confirms the :deep scaffolding
styles apply inside the children.
