+++ frontend/src/views/datastudio/components/DataStudioRightPanel.vue
@@ -29,445 +29,11 @@
 
         <el-tabs v-model="state.metaTab" class="meta-tabs detail-tabs">
           <el-tab-pane name="basic" label="基本信息">
-            <div class="meta-section meta-section-fill">
-              <div class="basic-grid" :class="{ single: !isDorisTable(state.table) }">
-                <section class="section-block">
-                  <div class="section-header">
-                    <div class="section-title">表信息</div>
-                    <div class="section-actions">
-                      <el-tooltip
-                        v-if="!state.metaEditing && isPlatformMetadataMissing(state.table)"
-                        content="请先同步到平台元数据后再操作"
-                        placement="top"
-                      >
-                        <span>
-                          <el-button type="primary" size="small" disabled>编辑</el-button>
-                        </span>
-                      </el-tooltip>
-                      <el-tooltip
-                        v-else-if="!state.metaEditing && isDorisTable(state.table) && !clusterId"
-                        content="请选择 Doris 集群后再编辑"
-                        placement="top"
-                      >
-                        <span>
-                          <el-button type="primary" size="small" disabled>编辑</el-button>
-                        </span>
-                      </el-tooltip>
-                      <el-button
-                        v-else-if="!state.metaEditing"
-                        type="primary"
-                        size="small"
-                        :disabled="isDemoMode"
-                        @click="startMetaEdit(activeTabId)"
-                      >
-                        编辑
-                      </el-button>
-
-                      <el-tooltip
-                        v-if="!state.metaEditing && isPlatformMetadataMissing(state.table)"
-                        content="请先同步到平台元数据后再操作"
-                        placement="top"
-                      >
-                        <span>
-                          <el-button type="danger" plain size="small" disabled>删除表</el-button>
-                        </span>
-                      </el-tooltip>
-                      <el-tooltip
-                        v-else-if="!state.metaEditing && isDorisTable(state.table) && !clusterId"
-                        content="请选择 Doris 集群后再删除"
-                        placement="top"
-                      >
-                        <span>
-                          <el-button type="danger" plain size="small" disabled>删除表</el-button>
-                        </span>
-                      </el-tooltip>
-                      <el-button
-                        v-else-if="!state.metaEditing"
-                        type="danger"
-                        plain
-                        size="small"
-                        :disabled="isDemoMode"
-                        @click="handleDeleteTable"
-                      >
-                        删除表
-                      </el-button>
-
-                      <template v-else>
-                        <el-button size="small" @click="cancelMetaEdit(activeTabId)">取消</el-button>
-                        <el-button
-                          type="primary"
-                          size="small"
-                          :loading="state.metaSaving"
-                          :disabled="isDemoMode"
-                          @click="saveMetaEdit(activeTabId)"
-                        >
-                          保存
-                        </el-button>
-                      </template>
-                    </div>
-                  </div>
-
-                  <el-scrollbar class="meta-scroll">
-                    <el-descriptions :column="1" border size="small" class="meta-descriptions">
-                      <el-descriptions-item label="表名">
-                        <el-input v-if="state.metaEditing" v-model="state.metaForm.tableName" size="small" class="meta-input" />
-                        <span v-else>{{ state.table.tableName || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="表注释">
-                        <el-input
-                          v-if="state.metaEditing"
-                          v-model="state.metaForm.tableComment"
-                          size="small"
-                          class="meta-input"
-                        />
-                        <span v-else>{{ state.table.tableComment || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="分层">
-                        <el-select
-                          v-if="state.metaEditing"
-                          v-model="state.metaForm.layer"
-                          size="small"
-                          placeholder="选择分层（必填）"
-                          class="meta-input"
-                        >
-                          <el-option v-for="item in layerOptions" :key="item.value" :label="item.label" :value="item.value" />
-                        </el-select>
-                        <span v-else>{{ state.table.layer || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="业务域">
-                        <el-select
-                          v-if="state.metaEditing"
-                          v-model="state.metaForm.businessDomain"
-                          size="small"
-                          placeholder="选择业务域"
-                          class="meta-input"
-                          @change="handleMetaBusinessDomainChange(activeTabId)"
-                        >
-                          <el-option
-                            v-for="item in businessDomainOptions"
-                            :key="item.domainCode"
-                            :label="`${item.domainCode} - ${item.domainName}`"
-                            :value="item.domainCode"
-                          />
-                        </el-select>
-                        <span v-else>{{ state.table.businessDomain || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="数据域">
-                        <el-select
-                          v-if="state.metaEditing"
-                          v-model="state.metaForm.dataDomain"
-                          size="small"
-                          placeholder="选择数据域"
-                          class="meta-input"
-                          :disabled="!state.metaForm.businessDomain"
-                        >
-                          <el-option
-                            v-for="item in dataDomainOptions"
-                            :key="item.domainCode"
-                            :label="`${item.domainCode} - ${item.domainName}`"
-                            :value="item.domainCode"
-                          />
-                        </el-select>
-                        <span v-else>{{ state.table.dataDomain || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="负责人">
-                        <el-input v-if="state.metaEditing" v-model="state.metaForm.owner" size="small" class="meta-input" />
-                        <span v-else>{{ state.table.owner || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="数据库">
-                        <span>{{ state.table.dbName || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="行数">
-                        <el-button
-                          link
-                          type="primary"
-                          class="metric-link"
-                          :disabled="!state.table?.id"
-                          @click="trendDialogRef?.open('rowCount')"
-                        >
-                          {{ formatRowCountDisplay(resolveTableRowCount(state.table)) }}
-                        </el-button>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="数据量">
-                        <el-button
-                          link
-                          type="primary"
-                          class="metric-link"
-                          :disabled="!state.table?.id"
-                          @click="trendDialogRef?.open('dataSize')"
-                        >
-                          {{ formatStorageSizeDisplay(resolveTableStorageSize(state.table)) }}
-                        </el-button>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="Doris创建时间">
-                        <span>{{ formatDateTime(resolveTableDorisCreateTime(state.table)) }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="Doris更新时间">
-                        <span>{{ formatDateTime(resolveTableDorisUpdateTime(state.table)) }}</span>
-                      </el-descriptions-item>
-                    </el-descriptions>
-                  </el-scrollbar>
-                </section>
-
-                <section v-if="isDorisTable(state.table)" class="section-block doris-block">
-                  <div class="section-header">
-                    <div class="section-title">Doris 配置</div>
-                    <el-tag size="small" type="warning" effect="plain">DORIS</el-tag>
-                  </div>
-
-                  <el-scrollbar class="meta-scroll">
-                    <el-descriptions :column="1" border size="small" class="meta-descriptions">
-                      <el-descriptions-item label="表模型">
-                        <span>{{ state.table.tableModel || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="主键列">
-                        <span>{{ state.table.keyColumns || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="分区字段">
-                        <span>{{ state.table.partitionColumn || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="分桶字段">
-                        <span>{{ state.table.distributionColumn || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="分桶数">
-                        <el-input-number
-                          v-if="state.metaEditing"
-                          v-model="state.metaForm.bucketNum"
-                          :min="1"
-                          size="small"
-                          controls-position="right"
-                          class="meta-input"
-                        />
-                        <span v-else>{{ state.table.bucketNum || '-' }}</span>
-                      </el-descriptions-item>
-                      <el-descriptions-item label="副本数">
-                        <template v-if="state.metaEditing">
-                          <div class="replica-edit">
-                            <el-input-number
-                              v-model="state.metaForm.replicaNum"
-                              :min="1"
-                              size="small"
-                              controls-position="right"
-                              class="meta-input"
-                            />
-                            <span v-if="isReplicaWarning(state.metaForm.replicaNum)" class="replica-warning">
-                              <el-icon><Warning /></el-icon>
-                              建议≥3
-                            </span>
-                          </div>
-                        </template>
-                        <span v-else :class="['replica-value', { 'replica-danger': isReplicaWarning(state.table.replicaNum) }]">
-                          <el-icon v-if="isReplicaWarning(state.table.replicaNum)" class="warning-icon"><Warning /></el-icon>
-                          {{ state.table.replicaNum || '-' }}
-                        </span>
-                      </el-descriptions-item>
-                    </el-descriptions>
-                  </el-scrollbar>
-                </section>
-              </div>
-            </div>
+            <DataStudioRightPanelBasic />
           </el-tab-pane>
 
           <el-tab-pane name="columns" label="列详情">
-            <div class="meta-section meta-section-fill">
-              <section class="section-block section-fill">
-                <div class="section-header">
-                  <div class="section-title">字段定义</div>
-                  <div class="section-actions">
-                    <el-tag
-                      v-if="state.fieldsEditing && isAggregateTable(state.table)"
-                      type="warning"
-                      size="small"
-                      effect="plain"
-                    >
-                      AGGREGATE 表仅支持修改注释
-                    </el-tag>
-                    <el-tag
-                      v-if="state.fieldsEditing && isDorisTable(state.table)"
-                      type="warning"
-                      size="small"
-                      effect="plain"
-                    >
-                      主键列不可在线修改
-                    </el-tag>
-
-                    <el-tooltip
-                      v-if="!state.fieldsEditing && isPlatformMetadataMissing(state.table)"
-                      content="请先同步到平台元数据后再操作"
-                      placement="top"
-                    >
-                      <span>
-                        <el-button type="primary" size="small" disabled>编辑</el-button>
-                      </span>
-                    </el-tooltip>
-                    <el-tooltip
-                      v-else-if="!state.fieldsEditing && isDorisTable(state.table) && !clusterId"
-                      content="请选择 Doris 集群后再编辑"
-                      placement="top"
-                    >
-                      <span>
-                        <el-button type="primary" size="small" disabled>编辑</el-button>
-                      </span>
-                    </el-tooltip>
-                    <el-button
-                      v-else-if="!state.fieldsEditing"
-                      type="primary"
-                      size="small"
-                      :disabled="isDemoMode"
-                      @click="startFieldsEdit(activeTabId)"
-                    >
-                      编辑
-                    </el-button>
-                    <template v-else>
-                      <el-button size="small" @click="cancelFieldsEdit(activeTabId)" :disabled="state.fieldSubmitting">
-                        取消
-                      </el-button>
-                      <el-button
-                        type="primary"
-                        size="small"
-                        :loading="state.fieldSubmitting"
-                        :disabled="isDemoMode"
-                        @click="saveFieldsEdit(activeTabId)"
-                      >
-                        保存修改
-                      </el-button>
-                    </template>
-                  </div>
-                </div>
-
-                <div v-if="fieldRows.length" class="meta-table">
-                  <el-table :data="fieldRows" border size="small" height="100%" class="columns-table">
-                    <el-table-column label="字段名" width="136" show-overflow-tooltip>
-                      <template #default="{ row }">
-                        <el-input
-                          v-if="state.fieldsEditing"
-                          v-model="row.fieldName"
-                          size="small"
-                          placeholder="字段名"
-                          :disabled="isAggregateTable(state.table)"
-                        />
-                        <span v-else>{{ row.fieldName }}</span>
-                      </template>
-                    </el-table-column>
-                    <el-table-column label="类型" width="136">
-                      <template #default="{ row }">
-                        <el-input
-                          v-if="state.fieldsEditing"
-                          v-model="row.fieldType"
-                          size="small"
-                          placeholder="VARCHAR(255)"
-                          :disabled="isAggregateTable(state.table)"
-                        />
-                        <span v-else>{{ row.fieldType }}</span>
-                      </template>
-                    </el-table-column>
-                    <el-table-column label="可为空" width="84">
-                      <template #default="{ row }">
-                        <el-switch
-                          v-if="state.fieldsEditing"
-                          v-model="row.isNullable"
-                          :active-value="1"
-                          :inactive-value="0"
-                          size="small"
-                          :disabled="isAggregateTable(state.table)"
-                        />
-                        <el-tag v-else :type="row.isNullable ? 'success' : 'danger'" size="small">
-                          {{ row.isNullable ? '是' : '否' }}
-                        </el-tag>
-                      </template>
-                    </el-table-column>
-                    <el-table-column label="主键" width="84">
-                      <template #default="{ row }">
-                        <template v-if="state.fieldsEditing">
-                          <el-tooltip v-if="isDorisTable(state.table)" content="Doris 不支持在线修改主键列" placement="top">
-                            <span>
-                              <el-switch v-model="row.isPrimary" :active-value="1" :inactive-value="0" size="small" disabled />
-                            </span>
-                          </el-tooltip>
-                          <el-switch
-                            v-else
-                            v-model="row.isPrimary"
-                            :active-value="1"
-                            :inactive-value="0"
-                            size="small"
-                            :disabled="isAggregateTable(state.table)"
-                          />
-                        </template>
-                        <el-tag v-else :type="row.isPrimary ? 'success' : 'info'" size="small">
-                          {{ row.isPrimary ? '是' : '否' }}
-                        </el-tag>
-                      </template>
-                    </el-table-column>
-                    <el-table-column label="默认值" width="120">
-                      <template #default="{ row }">
-                        <el-input
-                          v-if="state.fieldsEditing"
-                          v-model="row.defaultValue"
-                          size="small"
-                          placeholder="可选"
-                          :disabled="isAggregateTable(state.table)"
-                        />
-                        <span v-else>{{ row.defaultValue || '-' }}</span>
-                      </template>
-                    </el-table-column>
-                    <el-table-column label="注释" min-width="150" show-overflow-tooltip>
-                      <template #default="{ row }">
-                        <el-input v-if="state.fieldsEditing" v-model="row.fieldComment" size="small" placeholder="字段注释" />
-                        <span v-else>{{ row.fieldComment || '-' }}</span>
-                      </template>
-                    </el-table-column>
-                    <el-table-column v-if="state.fieldsEditing" label="操作" width="150" fixed="right">
-                      <template #default="{ row }">
-                        <el-tooltip
-                          v-if="isAggregateTable(state.table)"
-                          content="AGGREGATE 表不支持新增字段"
-                          placement="top"
-                        >
-                          <span>
-                            <el-button link type="primary" size="small" disabled>新增</el-button>
-                          </span>
-                        </el-tooltip>
-                        <el-button v-else link type="primary" size="small" @click="addField(activeTabId, row)">新增</el-button>
-                        <el-popconfirm
-                          width="240"
-                          confirm-button-text="确定"
-                          cancel-button-text="取消"
-                          :title="`确定删除字段「${row.fieldName || '未命名'}」吗？`"
-                          @confirm="removeField(activeTabId, row)"
-                        >
-                          <template #reference>
-                            <el-tooltip
-                              v-if="isAggregateTable(state.table)"
-                              content="AGGREGATE 表不支持删除字段"
-                              placement="top"
-                            >
-                              <span>
-                                <el-button link type="danger" size="small" disabled>删除</el-button>
-                              </span>
-                            </el-tooltip>
-                            <el-button v-else link type="danger" size="small">删除</el-button>
-                          </template>
-                        </el-popconfirm>
-                      </template>
-                    </el-table-column>
-                  </el-table>
-                </div>
-
-                <el-empty v-else description="暂无字段" :image-size="60">
-                  <template #default>
-                    <el-button
-                      v-if="state.fieldsEditing"
-                      type="primary"
-                      size="small"
-                      :disabled="isDemoMode || isAggregateTable(state.table)"
-                      @click="addField(activeTabId)"
-                    >
-                      新增字段
-                    </el-button>
-                  </template>
-                </el-empty>
-              </section>
-            </div>
+            <DataStudioRightPanelColumns />
           </el-tab-pane>
 
           <el-tab-pane name="ddl" label="DDL">
@@ -490,64 +56,7 @@
           </el-tab-pane>
 
           <el-tab-pane name="access" label="访问情况">
-            <div class="meta-section meta-section-fill" v-loading="state.accessLoading">
-              <section class="section-block section-fill">
-                <div class="section-header">
-                  <div class="section-title">访问概况</div>
-                  <div class="section-actions">
-                    <el-button size="small" :disabled="!state.table?.id || state.accessLoading" @click="refreshAccess">
-                      刷新
-                    </el-button>
-                  </div>
-                </div>
-
-                <el-scrollbar class="meta-scroll access-scroll">
-                  <template v-if="state.accessStats">
-                    <el-alert
-                      v-if="state.accessStats.note"
-                      :title="state.accessStats.note"
-                      type="warning"
-                      show-icon
-                      :closable="false"
-                      class="access-note"
-                    />
-
-                    <div class="metrics-grid">
-                      <div v-for="metric in accessMetrics" :key="metric.label" class="metric-card">
-                        <div class="metric-label">{{ metric.label }}</div>
-                        <div class="metric-value">{{ metric.value }}</div>
-                      </div>
-                    </div>
-
-                    <div class="section-divider"></div>
-
-                    <div class="section-header small">
-                      <span>近{{ state.accessStats.trendDays || 14 }}天访问趋势</span>
-                    </div>
-                    <el-table :data="state.accessStats.trend || []" border size="small" class="access-table">
-                      <el-table-column prop="date" label="日期" min-width="120" />
-                      <el-table-column prop="accessCount" label="访问次数" width="120" />
-                    </el-table>
-
-                    <div class="section-divider"></div>
-
-                    <div class="section-header small">
-                      <span>活跃用户 Top{{ (state.accessStats.topUsers || []).length }}</span>
-                    </div>
-                    <el-table :data="state.accessStats.topUsers || []" border size="small" class="access-table">
-                      <el-table-column prop="userId" label="用户" min-width="140" show-overflow-tooltip />
-                      <el-table-column prop="accessCount" label="访问次数" width="100" />
-                      <el-table-column label="最近访问" min-width="160">
-                        <template #default="{ row }">
-                          {{ formatDateTime(row.lastAccessTime) }}
-                        </template>
-                      </el-table-column>
-                    </el-table>
-                  </template>
-                  <el-empty v-else :description="state.accessError || '暂无访问数据'" :image-size="60" />
-                </el-scrollbar>
-              </section>
-            </div>
+            <DataStudioRightPanelAccess />
           </el-tab-pane>
 
           <el-tab-pane name="versions" label="版本" lazy>
@@ -583,8 +92,6 @@
         @create-task="(type) => goCreateRelatedTask(activeTabId, type)"
         @go-lineage="goLineage(activeTabId)"
       />
-
-      <TableTrendDialog ref="trendDialogRef" :table="state.table" />
     </div>
 
     <div v-else class="right-empty">
@@ -594,20 +101,13 @@
 </template>
 
 <script setup>
-import { computed, inject, ref } from 'vue'
-import { Warning } from '@element-plus/icons-vue'
+import { computed, inject } from 'vue'
 import DataStudioRightPanelLineage from './DataStudioRightPanelLineage.vue'
 import TableVersionHistoryPanel from './TableVersionHistoryPanel.vue'
-import TableTrendDialog from './TableTrendDialog.vue'
+import DataStudioRightPanelBasic from './DataStudioRightPanelBasic.vue'
+import DataStudioRightPanelColumns from './DataStudioRightPanelColumns.vue'
+import DataStudioRightPanelAccess from './DataStudioRightPanelAccess.vue'
 import { isDemoMode } from '@/demo/runtime'
-import {
-  resolveTableRowCount,
-  resolveTableStorageSize,
-  resolveTableDorisCreateTime,
-  resolveTableDorisUpdateTime,
-  formatRowCountDisplay,
-  formatStorageSizeDisplay,
-} from '../tableFormat'
 import { usePanelVerticalResize } from '../composables/usePanelVerticalResize'
 
 const props = defineProps({
@@ -624,33 +124,12 @@ if (!ctx) {
 }
 
 const {
-  clusterId,
   openTabs,
   activeTab,
   tabStates,
-  layerOptions,
-  businessDomainOptions,
-  getMetaDataDomainOptions,
-  handleMetaBusinessDomainChange,
-  isDorisTable,
   isPlatformMetadataMissing,
-  isAggregateTable,
-  isReplicaWarning,
-  getFieldRows,
-  startMetaEdit,
-  cancelMetaEdit,
-  saveMetaEdit,
-  handleDeleteTable,
   syncMissingTableMetadata,
-  startFieldsEdit,
-  cancelFieldsEdit,
-  saveFieldsEdit,
-  addField,
-  removeField,
   copyDdl,
-  loadAccessStats,
-  formatDuration,
-  formatDateTime,
   goLineage,
   goCreateRelatedTask,
   openTask,
@@ -664,8 +143,6 @@ const activeTabItem = computed(() => {
   if (!id) return null
   return (openTabs.value || []).find((item) => String(item?.id) === id) || null
 })
-
-const trendDialogRef = ref(null)
 
 const rootClass = computed(() => [
   'right-root',
@@ -699,32 +176,6 @@ const {
   hasTableTab,
 })
 
-const fieldRows = computed(() => getFieldRows(activeTabId.value))
-const dataDomainOptions = computed(() => getMetaDataDomainOptions(activeTabId.value))
-
-const accessMetrics = computed(() => {
-  const stats = state.value?.accessStats
-  if (!stats) return []
-  return [
-    { label: '总访问次数', value: stats.totalAccessCount ?? 0 },
-    { label: `最近${stats.recentDays || 30}天`, value: stats.recentAccessCount ?? 0 },
-    { label: '访问用户数', value: stats.distinctUserCount ?? 0 },
-    { label: '平均耗时', value: formatAccessDuration(stats.averageDurationMs) },
-    { label: '最近访问', value: formatDateTime(stats.lastAccessTime) },
-    { label: '审计来源', value: stats.dorisAuditEnabled ? (stats.dorisAuditSource || '已启用') : '未启用' }
-  ]
-})
-
-const refreshAccess = () => {
-  const tabId = activeTabId.value
-  if (!tabId) return
-  loadAccessStats(tabId, true)
-}
-
-const formatAccessDuration = (value) => {
-  if (value === null || value === undefined || value === '') return '-'
-  return formatDuration(Number(value))
-}
 </script>
 
 <style scoped>
@@ -1030,30 +481,18 @@ const formatAccessDuration = (value) => {
   min-height: 0;
 }
 
-.meta-section {
+.meta-tabs :deep(.meta-section) {
   display: flex;
   flex-direction: column;
   gap: 10px;
 }
 
-.meta-section-fill {
+.meta-tabs :deep(.meta-section-fill) {
   flex: 1;
   min-height: 0;
 }
 
-.basic-grid {
-  flex: 1;
-  min-height: 0;
-  display: grid;
-  grid-template-columns: minmax(0, 1fr) minmax(0, 0.9fr);
-  gap: 10px;
-}
-
-.basic-grid.single {
-  grid-template-columns: 1fr;
-}
-
-.section-block {
+.meta-tabs :deep(.section-block) {
   border: 1px solid var(--line);
   border-radius: 8px;
   background: var(--panel-muted);
@@ -1064,120 +503,43 @@ const formatAccessDuration = (value) => {
   min-height: 0;
 }
 
-.section-fill {
+.meta-tabs :deep(.section-fill) {
   flex: 1;
 }
 
-.doris-block {
-  background: var(--accent-soft);
-}
-
-.section-header {
+.meta-tabs :deep(.section-header) {
   display: flex;
   align-items: center;
   justify-content: space-between;
   gap: 8px;
 }
 
-.section-title {
+.meta-tabs :deep(.section-title) {
   font-size: 13px;
   font-weight: 700;
   color: var(--text);
 }
 
-.section-actions {
+.meta-tabs :deep(.section-actions) {
   display: inline-flex;
   align-items: center;
   gap: 8px;
   flex-wrap: wrap;
 }
 
-.section-header.small {
-  font-size: 12px;
-  color: var(--text-sub);
-}
-
-.section-divider {
-  height: 1px;
-  margin: 10px 0;
-  background: var(--line);
-}
-
-.meta-scroll {
+.meta-tabs :deep(.meta-scroll) {
   flex: 1;
   min-height: 0;
   max-height: 100%;
   overflow: auto;
 }
 
-.meta-scroll :deep(.el-scrollbar__view) {
+.meta-tabs :deep(.meta-scroll .el-scrollbar__view) {
   padding-right: 4px;
   box-sizing: border-box;
 }
 
-.meta-descriptions :deep(.el-descriptions__label.is-bordered-label) {
-  width: 108px;
-  min-width: 108px;
-  white-space: nowrap;
-  color: var(--text-sub);
-}
-
-.meta-descriptions :deep(.el-descriptions__content.is-bordered-content) {
-  white-space: normal;
-  word-break: break-word;
-  overflow-wrap: anywhere;
-}
-
-.meta-input {
-  width: 100%;
-}
-
-.metric-link {
-  padding: 0;
-  font-weight: 600;
-}
-
-.replica-edit {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  flex-wrap: wrap;
-}
-
-.replica-warning {
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-  font-size: 12px;
-  color: #d14343;
-}
-
-.replica-value {
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-}
-
-.replica-danger {
-  color: #d14343;
-  font-weight: 600;
-}
-
-.warning-icon {
-  font-size: 12px;
-}
-
-.meta-table {
-  flex: 1;
-  min-height: 0;
-}
-
 :deep(.columns-table th.el-table__cell),
-:deep(.access-table th.el-table__cell) {
-  background: #f2f7ff;
-  color: var(--text-sub);
-}
-
 .code-shell {
   flex: 1;
   min-height: 0;
@@ -1204,41 +566,6 @@ const formatAccessDuration = (value) => {
   padding: 10px 12px;
   font-size: 12px;
   color: var(--text-muted);
-}
-
-.access-note {
-  margin-bottom: 10px;
-}
-
-.access-scroll {
-  flex: 1;
-  min-height: 0;
-}
-
-.metrics-grid {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  gap: 8px;
-}
-
-.metric-card {
-  border: 1px solid var(--line);
-  border-radius: 8px;
-  background: #fff;
-  padding: 8px 9px;
-}
-
-.metric-label {
-  font-size: 11px;
-  color: var(--text-sub);
-}
-
-.metric-value {
-  margin-top: 4px;
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--text);
-  word-break: break-word;
 }
 
 .lineage-panel {
